fix: stabilize Cloudflare preview requests
