+++ src/components/landing/PublicSiteHeader.tsx
@@ -48,8 +48,10 @@ function MenuRow({
   isLast: boolean;
   onClick: () => void;
 }) {
+  const MenuLink = href.includes("#") ? "a" : Link;
+
   return (
-    <Link
+    <MenuLink
       href={href}
       onClick={onClick}
       className={cn(
@@ -76,7 +78,7 @@ function MenuRow({
           {description}
         </span>
       </span>
-    </Link>
+    </MenuLink>
   );
 }
 
@@ -198,6 +200,7 @@ export function PublicSiteHeader({
       <nav className="mx-auto flex h-16 w-full max-w-6xl items-center justify-between px-6">
         <Link
           href="/"
+          prefetch={false}
           aria-label={t("header.logoAriaLabel")}
           className="flex items-center"
         >
@@ -211,9 +214,10 @@ export function PublicSiteHeader({
             const isActive =
               (currentPage === "home" && item.key === "home") ||
               (currentPage === "gallery" && item.key === "gallery");
+            const NavLink = href.includes("#") ? "a" : Link;
 
             return (
-              <Link
+              <NavLink
                 key={item.key}
                 href={href}
                 className={cn(
@@ -222,7 +226,7 @@ export function PublicSiteHeader({
                 )}
               >
                 {item.label}
-              </Link>
+              </NavLink>
             );
           })}
         </div>
@@ -237,6 +241,7 @@ export function PublicSiteHeader({
             </div>
             <Link
               href="/studio"
+              prefetch={false}
               className="inline-flex h-9 items-center gap-1.5 rounded-full bg-[#1E93DB] px-4 text-sm font-medium text-white shadow-md shadow-[#1E93DB]/30 transition hover:brightness-110"
             >
               {t("header.openStudio")} <ArrowRight className="size-3.5" />
@@ -320,6 +325,7 @@ export function PublicSiteHeader({
 
                     <Link
                       href="/studio"
+                      prefetch={false}
                       onClick={() => setOpenMobileMenu(false)}
                       className="mt-3 inline-flex h-10 w-full items-center justify-center gap-2 rounded-full bg-[#1E93DB] px-4 text-sm font-medium text-white shadow-md shadow-[#1E93DB]/25 transition hover:brightness-110"
                     >
